FD 管理レコードと事業所数レコードの共通化
