+++ lib/kirico/models/base_company_count.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+require 'virtus'
+require 'active_model'
+
+module Kirico
+  class BaseCompanyCount
+    include Virtus.model
+    include ActiveModel::Validations
+  end
+end
