+++ lib/kirico/models/sr_company_count.rb
@@ -1,22 +1,14 @@
 # frozen_string_literal: true
-require 'virtus'
-require 'active_model'
+require 'kirico/models/base_company_count'
 
 # 社会保険労務士: 事業所数情報
 module Kirico
-  class SrCompanyCount
-    include Virtus.model
-    include ActiveModel::Validations
-
+  class SrCompanyCount < BaseCompanyCount
     COUNT = '001'
 
     attribute :name, String
 
     validates :name, charset: { accept: [:katakana, :kanji] }, sjis_bytesize: { in: 1..25 }
-
-    def initialize
-      super
-    end
 
     def to_csv
       [name, COUNT].map { |attr| attr.to_s.encode('CP932') }.join(',')
